+++ jest.config.js
@@ -1,5 +1,10 @@
+const files = `<rootDir>/test/unit/modules/*${
+  process.env.TEST_ENV ? `-${process.env.TEST_ENV}-` : ``
+}*.js`;
+
 module.exports = {
   verbose: true,
   bail: true,
-  testMatch: ["<rootDir>/test/unit/modules/*.js"]
+  testMatch: [files],
+  testEnvironment: process.env.TEST_ENV || 'jsdom'
 };