+++ test/unit/modules/browser-auto-tracking-spec.js
@@ -56,13 +56,6 @@ describe('Auto Tracking', () => {
       timestamp: expect.any(String),
       addons: [
         {
-          name: 'keen:ip_to_geo',
-          input: {
-            ip: 'ip_address'
-          },
-          output : 'geo'
-        },
-        {
           name: 'keen:ua_parser',
           input: {
             ua_string: 'user_agent'
@@ -96,6 +89,13 @@ describe('Auto Tracking', () => {
             date_time: 'local_time_full'
           },
           output: 'time.local'
+        },
+        {
+          name: 'keen:ip_to_geo',
+          input: {
+            ip: 'ip_address'
+          },
+          output : 'geo'
         }
       ],
     }
@@ -136,13 +136,6 @@ describe('Auto Tracking', () => {
       timestamp: expect.any(String),
       addons: [
         {
-          name: 'keen:ip_to_geo',
-          input: {
-            ip: 'ip_address'
-          },
-          output : 'geo'
-        },
-        {
           name: 'keen:ua_parser',
           input: {
             ua_string: 'user_agent'
@@ -176,6 +169,13 @@ describe('Auto Tracking', () => {
             date_time: 'local_time_full'
           },
           output: 'time.local'
+        },
+        {
+          name: 'keen:ip_to_geo',
+          input: {
+            ip: 'ip_address'
+          },
+          output : 'geo'
         }
       ],
     },
@@ -197,8 +197,6 @@ describe('Auto Tracking', () => {
       description: expect.any(String),
       time_on_page: expect.any(Number)
     },
-    ip_address: '${keen.ip}',
-    geo: { },
     user_agent: '${keen.user_agent}',
     tech: {
       profile: expect.any(Object)
@@ -218,13 +216,6 @@ describe('Auto Tracking', () => {
     keen: {
       timestamp: expect.any(String),
       addons: [
-        {
-          name: 'keen:ip_to_geo',
-          input: {
-            ip: 'ip_address'
-          },
-          output : 'geo'
-        },
         {
           name: 'keen:ua_parser',
           input: {
@@ -259,6 +250,13 @@ describe('Auto Tracking', () => {
             date_time: 'local_time_full'
           },
           output: 'time.local'
+        },
+        {
+          name: 'keen:ip_to_geo',
+          input: {
+            ip: 'ip_address'
+          },
+          output : 'geo'
         }
       ],
     },
@@ -327,6 +325,24 @@ describe('Auto Tracking', () => {
     const uuid = cookie.get('uuid');
     expect(uuid).not.toBe(null);
     expect(uuid.length).toBeGreaterThan(0);
+  });
+
+  describe('GDPR', () => {
+    beforeEach(() => {
+      client1 = new Keen({
+        projectId: 'aa',
+        writeKey: 'bb'
+      });
+      client1.on('recordEvent', mockFn1);
+      mockFn1.mockClear();
+      client1.initAutoTracking({
+        collectIpAddress: false
+      });
+    });
+
+    it('should not collect IP addresses', () => {
+      expect(mockFn1.mock.calls[0][1].ip_address).toEqual(undefined);
+    });
   });
 
 });